Ejercicio 4 corregido

+++ 04_area_triangulo/Ejercicio4/src/Ejercicio4.java
@@ -6,10 +6,10 @@ public class Ejercicio4 {
     public static void main(String[] args) throws IOException{
         BufferedReader br=new BufferedReader(new InputStreamReader(System.in));
         System.out.println("Escriba el valor de la base");
-        int b=Integer.parseInt(br.readLine());
+        double b=Double.parseDouble(br.readLine());
         System.out.println("Escriba el valor de la altura");
-        int h=Integer.parseInt(br.readLine());
-        int a=(b*h)/2;
+        double h=Double.parseDouble(br.readLine());
+        double a=(b*h)/2;
         System.out.println("El area del triángulo es: "+a);
